feat: extend conflict detection to handle subdomain globs (#14)
Some users want to set records for both sub.mydomain.com and *.sub.mydomain.com. They might try to do so using two commands:

$ netlify-ddns -d mydomain.com -s 'sub' --ttl 3600 -t <token>
$ netlify-ddns -d mydomain.com -s '*.sub' --ttl 3600 -t <token>

However, this was resulting in the glob subdomains never being deleted, because the existing Netlify DNS records were not being detected as duplicates. This was because the assumption of only 3 URL segments was baked into the code that filters for conflicts.

To handle this case, this commit out what the entire target hostname should be for the request, and compares the Netlify hostnames against that instead of breaking it up into segments. I also broke the conflict detection out into a separate function and added some tests that ensure the conflict/exact match detection still works as expected.

As a side effect, this handles the TODO item about empty subdomains.